[FIX] im_livechat: fix unknown models warning
The chatbot relies on several models to work on embed pages
such as: `chatbot.script` or `chatbot.script.step`. Data related
to those models are sent to the store when a chatbot is active on
a thread. However, those JS models are only preset in the live
chat embed bundle. As a result, warning are emitted on the odoo
backend when data related to those models are received, and ignore.
This PR moves the models to the common bundle. While not ideal, this
approach will allow the live chat tours to be executed as warning are
making them fail.
