+++ addons/im_livechat/static/src/core/common/chatbot_model.js
