+++ addons/im_livechat/static/src/core/common/chatbot_script_step_model.js
