+++ addons/im_livechat/static/src/core/common/chatbot_step_model.js
