+++ addons/im_livechat/static/src/core/common/message_model_patch.js
@@ -1,9 +1,14 @@
 import { Message } from "@mail/core/common/message_model";
+import { Record } from "@mail/core/common/record";
 
 import { patch } from "@web/core/utils/patch";
 
 /** @type {import("models").Message} */
 const messagePatch = {
+    setup() {
+        super.setup(...arguments);
+        this.chatbotStep = Record.one("ChatbotStep", { inverse: "message" });
+    },
     isTranslatable(thread) {
         return (
             super.isTranslatable(thread) ||
