+++ addons/im_livechat/static/src/embed/common/message_model_patch.js
@@ -1,15 +1,10 @@
 import { Message } from "@mail/core/common/message_model";
-import { Record } from "@mail/core/common/record";
 
 import { patch } from "@web/core/utils/patch";
 import { SESSION_STATE } from "./livechat_service";
 
 /** @type {import("models").Message} */
 const messagePatch = {
-    setup() {
-        super.setup();
-        this.chatbotStep = Record.one("ChatbotStep", { inverse: "message" });
-    },
     canAddReaction(thread) {
         return (
             super.canAddReaction(thread) &&
