+++ addons/im_livechat/static/src/embed/common/message_patch.js
@@ -16,7 +16,7 @@ patch(Message.prototype, {
     },
 
     /**
-     * @param {import("@im_livechat/embed/common/chatbot/chatbot_step_model").StepAnswer} answer
+     * @param {import("@im_livechat/core/common/chatbot_step_model").StepAnswer} answer
      */
     answerChatbot(answer) {
         return this.props.message.thread.post(answer.name, {}, { selected_answer_id: answer.id });
